each test has own NumCompressNodes
