+++ tests/post_training/conftest.py
@@ -9,6 +9,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from pathlib import Path
+
+import pytest
+
 
 def pytest_addoption(parser):
     parser.addoption("--data", action="store", help="Data directory")
@@ -35,3 +39,66 @@ def pytest_addoption(parser):
         help="Report memory using MemoryMonitor from tools/memory_monitor.py. "
         "Warning: currently, reported memory values are not always reproducible.",
     )
+
+
+@pytest.fixture(scope="session", name="data_dir")
+def fixture_data(pytestconfig):
+    if pytestconfig.getoption("data") is None:
+        msg = "This test requires the --data argument to be specified."
+        raise ValueError(msg)
+    return Path(pytestconfig.getoption("data"))
+
+
+@pytest.fixture(scope="session", name="output_dir")
+def fixture_output(pytestconfig):
+    return Path(pytestconfig.getoption("output"))
+
+
+@pytest.fixture(scope="session", name="no_eval")
+def fixture_no_eval(pytestconfig):
+    return pytestconfig.getoption("no_eval")
+
+
+@pytest.fixture(scope="session", name="batch_size")
+def fixture_batch_size(pytestconfig):
+    return pytestconfig.getoption("batch_size")
+
+
+@pytest.fixture(scope="session", name="subset_size")
+def fixture_subset_size(pytestconfig):
+    return pytestconfig.getoption("subset_size")
+
+
+@pytest.fixture(scope="session", name="run_fp32_backend")
+def fixture_run_fp32_backend(pytestconfig):
+    return pytestconfig.getoption("fp32")
+
+
+@pytest.fixture(scope="session", name="run_torch_cuda_backend")
+def fixture_run_torch_cuda_backend(pytestconfig):
+    return pytestconfig.getoption("cuda")
+
+
+@pytest.fixture(scope="session", name="run_benchmark_app")
+def fixture_run_benchmark_app(pytestconfig):
+    return pytestconfig.getoption("benchmark")
+
+
+@pytest.fixture(scope="session", name="torch_compile_validation")
+def fixture_torch_compile_validation(pytestconfig):
+    return pytestconfig.getoption("torch_compile_validation")
+
+
+@pytest.fixture(scope="session", name="extra_columns")
+def fixture_extra_columns(pytestconfig):
+    return pytestconfig.getoption("extra_columns")
+
+
+@pytest.fixture(scope="session", name="memory_monitor")
+def fixture_memory_monitor(pytestconfig):
+    return pytestconfig.getoption("memory_monitor")
+
+
+@pytest.fixture(scope="session", name="forked")
+def fixture_forked(pytestconfig):
+    return pytestconfig.getoption("forked")
